% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatabias.r
\name{moredatabias}
\alias{moredatabias}
\title{moredatabias}
\usage{
moredatabias(
  datfile.origsave,
  dat_list,
  locnum,
  obsnum,
  betavar,
  uparams,
  abundse = NULL,
  catchscale = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatabias
}
