% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatabiasreadrel.r
\name{moredatabiasreadrel}
\alias{moredatabiasreadrel}
\title{moredatabiasreadrel}
\usage{
moredatabiasreadrel(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatabiasreadrel
}
