% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatabiasreadrelSE.r
\name{moredatabiasreadrelSE}
\alias{moredatabiasreadrelSE}
\title{moredatabiasreadrelSE}
\usage{
moredatabiasreadrelSE(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatabiasreadrelSE
}
