% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatabiasrel.r
\name{moredatabiasrel}
\alias{moredatabiasrel}
\title{moredata}
\usage{
moredatabiasrel(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredata
}
