% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatadahl.r
\name{moredatadahl}
\alias{moredatadahl}
\title{moredatadahl}
\usage{
moredatadahl(datfile.origsave, dat_list, locnum, obsnum, betavar, uparams, ...)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatadahl
}
