% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatadahlread.r
\name{moredatadahlread}
\alias{moredatadahlread}
\title{moredata}
\usage{
moredatadahlread(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredata
}
