% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatadahlreadrel.r
\name{moredatadahlreadrel}
\alias{moredatadahlreadrel}
\title{moredata}
\usage{
moredatadahlreadrel(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredata
}
