% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatadahlreadrelSE.r
\name{moredatadahlreadrelSE}
\alias{moredatadahlreadrelSE}
\title{moredatadahlreadrelSE}
\usage{
moredatadahlreadrelSE(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatadahlreadrelSE
}
