% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatadahlrel.r
\name{moredatadahlrel}
\alias{moredatadahlrel}
\title{moredatadahlrel}
\usage{
moredatadahlrel(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatadahlrel
}
