% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatadahlrelSE.r
\name{moredatadahlrelSE}
\alias{moredatadahlrelSE}
\title{moredatadahlrelSE}
\usage{
moredatadahlrelSE(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatadahlrelSE
}
