% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatadahlSE.r
\name{moredatadahlSE}
\alias{moredatadahlSE}
\title{moredatadahlSE}
\usage{
moredatadahlSE(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatadahlSE
}
