% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredataeffort.r
\name{moredataeffort}
\alias{moredataeffort}
\title{moredata}
\usage{
moredataeffort(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredata
}
