% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatasample.r
\name{moredatasample}
\alias{moredatasample}
\title{moredatasample}
\usage{
moredatasample(
  datfile.origsave,
  dat_list,
  locnum,
  obsnum,
  betavar,
  uparams,
  abundse = NULL,
  abundscale = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatasample
}
