% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatasamplerel.r
\name{moredatasamplerel}
\alias{moredatasamplerel}
\title{moredata}
\usage{
moredatasamplerel(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredata
}
