% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatasamplerelSE.r
\name{moredatasamplerelSE}
\alias{moredatasamplerelSE}
\title{moredatasamplerelSE}
\usage{
moredatasamplerelSE(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatasamplerelSE
}
