% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatatest_standard.r
\name{moredatatest_standard}
\alias{moredatatest_standard}
\title{moredatatest}
\usage{
moredatatest_standard(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredatatest
}
