% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fishery.r
\name{spatial_fishery}
\alias{spatial_fishery}
\title{spatial_fishery}
\usage{
spatial_fishery(
  locnum,
  obsnum,
  betavar,
  uparams,
  totcatches,
  year,
  random = FALSE,
  avghauls
)
}
\arguments{
\item{locnum}{description}

\item{obsnum}{description}
}
\description{
spatial_fishery
}
