% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatabias.r, R/moredatabiasSE.r
\name{moredatabias}
\alias{moredatabias}
\title{moredata}
\usage{
moredatabias(datfile.origsave, dat_list)

moredatabias(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredata

moredata
}
