% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moredatasample.r
\name{moredatasample}
\alias{moredatasample}
\title{moredata}
\usage{
moredatasample(datfile.origsave, dat_list)
}
\arguments{
\item{datfile.origsave}{description}

\item{dat_list}{description}
}
\description{
moredata
}
